Refactor: common AbstractTelemetryIT
