+++ test/external-modules/apm-integration/src/javaRestTest/java/org/elasticsearch/test/apmintegration/OtelMetricsIT.java
@@ -36,7 +36,7 @@ public class OtelMetricsIT extends AbstractMetricsIT {
         .build();
 
     @ClassRule
-    public static TestRule ruleChain = AbstractMetricsIT.buildRuleChain(recordingApmServer, cluster);
+    public static TestRule ruleChain = buildRuleChain(recordingApmServer, cluster);
 
     @Override
     protected RecordingApmServer apmServer() {
